refactor: id값 int->long

+++ app/src/main/java/woowacourse/movie/data/dao/TicketDao.kt
@@ -27,7 +27,7 @@ interface TicketDao {
         seats: List<SeatEntity>,
     ) {
         val ticketId = save(ticket)
-        save(seats.map { it.copy(ticketId = ticketId.toInt()) })
+        save(seats.map { it.copy(ticketId = ticketId) })
     }
 }
